refactor: __init__.py files usage changes
