--- Config/__init__.py
@@ -1,1 +1,0 @@
-from .config import API_TOKEN
