+++ Controllers/UserOptionsControllers/user_options_controller.py
@@ -4,8 +4,8 @@ from aiogram.fsm.context import FSMContext
 from aiogram.types import Message, CallbackQuery
 
 from Controllers.UserOptionsControllers.help_functions import delete_message
-from Utils.FSM.UserOptions import UserOptionsFSM
-from Utils.Keyboards.Inline.UserOptions import (get_user_options_menu, get_user_add_menu)
+from Utils.FSM.UserOptions.UserOptionsFSM import UserOptionsFSM
+from Utils.Keyboards.Inline.UserOptions.user_options_keyboards import get_user_options_menu, get_user_add_menu
 
 
 async def open_user_option_menu_controller(message: Message, state: FSMContext):
