--- Controllers/__init__.py
@@ -1,13 +1,0 @@
-from .command_controller import start_controller, help_controller, stop_controller
-from Controllers.UserOptionsControllers.user_options_controller import (open_user_option_menu_controller,
-                                                                        close_user_option_menu_controller,
-                                                                        open_user_add_menu_controller, )
-
-from Controllers.UserOptionsControllers.user_add_controller import (get_wait_user_id_message_controller,
-                                                                    get_wait_user_name_message_controller,
-                                                                    get_wait_user_default_payment_message_controller,
-                                                                    write_new_user_id_controller,
-                                                                    write_new_user_name_controller,
-                                                                    add_new_user_main_cancel_controller,
-                                                                    add_new_user_cancel_controller,
-write_new_user_default_payment_controller, add_new_user_controller)
