+++ Models/MonthPayment.py
@@ -1,7 +1,7 @@
 from peewee import *
 
-from Models import BaseModel
-from Models import User
+from Models.BaseModel import BaseModel
+from Models.User import User
 
 
 class MonthPayment(BaseModel):
