--- Models/__init__.py
@@ -1,4 +1,0 @@
-from .BaseModel import BaseModel, db
-from .User import User
-from .Debit import Debit
-from .MonthPayment import MonthPayment
