--- Utils/FSM/UserOptions/__init__.py
@@ -1,1 +1,0 @@
-from .UserOptionsFSM import UserOptionsFSM
