--- Utils/Keyboards/Inline/UserOptions/__init__.py
@@ -1,3 +1,0 @@
-from .const import (USER_OPTIONS_MENU_BUTTONS, USER_ADD_MENU_BUTTONS, USER_CANCEL_MENU, USER_CHANGE_MENU,
-                    BUTTONS_IN_FIRST_LINE)
-from .user_options_keyboards import get_user_options_menu, get_user_add_menu, get_user_cancel_menu, get_user_change_menu
